+++ pkg/template/master_template.go
@@ -753,7 +753,7 @@ storage:
       filesystem: root
       mode: 0644
       contents:
-        source: "data:text/plain;charset=utf-8;base64,{{  index .Files "conf/crictl" }}"
+        source: "data:text/plain;charset=utf-8;base64,{{  index .Files "conf/crictl.yaml" }}"
 
     - path: /etc/profile.d/setup-etcdctl.sh
       filesystem: root