+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+### Fixed
+
+- Copy `crictl.yaml` on master and worker nodes.
+
 ## [14.0.0] - 2022-07-14
 
 ### Changed